files

+++ logger.py
@@ -93,10 +93,13 @@ class VisdomLogger(BaseLogger):
         self.port = kwargs.pop('port', 7000)
         self.server = kwargs.pop('server', '35.229.22.191')
         self.env = kwargs.pop('env', 'main')
+        self.delete = kwargs.pop('delete', True)
+        print ("No deletion")
         print ("In (git) scaling-reset")
         print (f"Logging to environment {self.env}")
         self.visdom = visdom.Visdom(server="http://" + self.server, port=self.port, env=self.env)
-        self.visdom.delete_env(self.env)
+        if self.delete:
+            self.visdom.delete_env(self.env)
         self.windows = {}
         super().__init__(*args, **kwargs)
         self.save()
